+++ Sources/Scaletor/Models/MajorMode.swift
@@ -1,9 +1,11 @@
 import Foundation
 
-public enum Mode: String, CaseIterable {
+public enum MajorMode: String, CaseIterable {
     case ionian, dorian, phrygian, lydian, mixolydian, aeolian, locrian
+}
 
-    public var chords: [Chord] {
+extension MajorMode: Mode {
+    public var chords: [ChordVoicing] {
         [.major, .minor, .minor, .major, .major, .minor, .diminished]
             .offset(by: index)
     }