+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2020-07-26
+
+### Added
+
+- `Mode` protocol
+- Harmonic Minor modes (fixes #1)
+- Melodic Minor modes
+
+### Breaking
+
+- Rename `Chord` to `ChordVoicing`
+- Rename `Mode` to `MajorMode`
+
 ## [1.0.0] - 2020-07-11
 
 ### Added